Fix cross build
